requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.4
django-appconf==1.0.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-classy-tags==0.8.0
django-crispy-forms==1.7.2
django-easy-maps==0.9.3
django-geoposition==0.3.0
django-heroku==0.3.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
geographiclib==1.49
geopy==1.18.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.16.0
pandas==0.23.4
Pillow==5.4.1
psycopg2==2.7.6.1
pycodestyle==2.4.0
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
six==1.12.0
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wincertstore==0.2
wrapt==1.11.1